+++ Malfunctions/Patches/StartOfRoundPatches.cs
@@ -1,47 +1,49 @@
-﻿
-using System;
+﻿using System;
 using System.Linq;
-
+using Dissonance;
 using HarmonyLib;
-
 using Malfunctions.Helpers;
 
 namespace Malfunctions.Patches
 {
-	[HarmonyPatch(typeof(StartOfRound))]
-	public class StartOfRoundPatches
+    [HarmonyPatch(typeof(StartOfRound))]
+    public class StartOfRoundPatches
     {
         private static readonly System.Random rand = new System.Random();
 
-        public static bool malfunctionNavigation = false; // Whether the navigation malfunction is active.
-
         [HarmonyPostfix]
         [HarmonyPatch("EndOfGame")]
-        public static void TriggerMapMalfunction(StartOfRound __instance) {
-            MalfunctionsBase.Log.LogDebug(
-                String.Format(
-                    "Days left: {0}",
-                    TimeOfDay.Instance.daysUntilDeadline
-                )
-            );
+        public static void TriggerMapMalfunction(StartOfRound __instance)
+        {
+            Plugin.logger.LogDebug($"Days left: {TimeOfDay.Instance.daysUntilDeadline}");
 
             // If we previously had a malfunction make sure to unset it.
-            if(malfunctionNavigation || TimeOfDay.Instance.daysUntilDeadline < 2) {
-                malfunctionNavigation = false;
-
-			} else {
+            if (State.malfunctionNavigation || TimeOfDay.Instance.daysUntilDeadline < 2)
+            {
+                State.malfunctionNavigation = false;
+            }
+            else
+            {
                 // Make sure we can't trigger the navigation malfunction on the last day.
-                if (__instance.currentLevel.name != "CompanyBuildingLevel" && TimeOfDay.Instance.daysUntilDeadline >= 2) {
-
+                if (
+                    __instance.currentLevel.name != "CompanyBuildingLevel"
+                    && TimeOfDay.Instance.daysUntilDeadline >= 2
+                )
+                {
                     // Chance of navigation malfunctioning.
-                    double chance = MalfunctionsBase.ConfigMalfunctionNavigationChance.Value;
+                    double chance = Config.MalfunctionNavigationChance.Value;
                     double roll = rand.NextDouble() * 100;
 
-                    MalfunctionsBase.Log.LogDebug(String.Format("Roll: {0} / {1} ({2})", roll, chance, roll < chance ? "SUCCESS" : "FAIL"));
-  
-                    if (chance != 0 && roll < chance) {
+                    Plugin.logger.LogDebug(
+                        $"Roll: {roll} / {chance} ({(roll < chance ? "SUCCESS" : "FAIL")})"
+                    );
+
+                    if (chance != 0 && roll < chance)
+                    {
                         // Create an array of levels that we can use to filter out ones like the company.
-                        SelectableLevel[] filteredLevels = __instance.levels.Where(level => level.name != "CompanyBuildingLevel").ToArray();
+                        SelectableLevel[] filteredLevels = __instance
+                            .levels.Where(level => level.name != "CompanyBuildingLevel")
+                            .ToArray();
 
                         // Select the next random level from the list of all levels.
                         SelectableLevel selected = rand.NextFromCollection(__instance.levels);
@@ -49,19 +51,16 @@ namespace Malfunctions.Patches
                         // Set the current level to the one we have now selected.
                         __instance.ChangeLevel(selected.levelID);
 
-                        foreach (SelectableLevel level in filteredLevels) {
-                            MalfunctionsBase.Log.LogDebug(
-                                String.Format(
-                                    "Level ID: {0} / Level Name: {1}",
-                                    level.levelID,
-                                    level.name
-                                )
+                        foreach (SelectableLevel level in filteredLevels)
+                        {
+                            Plugin.logger.LogDebug(
+                                $"Level ID: {level.levelID} / Level Name: {level.name}"
                             );
                         }
 
-                        MalfunctionsBase.Log.LogDebug(String.Format("Selected Level ID: {0}", selected.levelID));
+                        Plugin.logger.LogDebug($"Selected Level ID: {selected.levelID}");
 
-                        malfunctionNavigation = true;
+                        State.malfunctionNavigation = true;
                     }
                 }
             }
@@ -69,22 +68,69 @@ namespace Malfunctions.Patches
 
         [HarmonyPostfix]
         [HarmonyPatch("SetMapScreenInfoToCurrentLevel")]
-        public static void HideMapScreenInfo(StartOfRound __instance) {
-            if (__instance.currentLevel.name != "CompanyBuildingLevel" && malfunctionNavigation == true) {
+        public static void HandleMalfunction(StartOfRound __instance)
+        {
+            if (
+                __instance.currentLevel.name != "CompanyBuildingLevel"
+                && State.malfunctionNavigation == true
+            )
+            {
                 // TODO: Swap this out with a custom HUD animation prefab similar to the radation one.
-                HUDManager.Instance.globalNotificationText.text = "SHIP MALFUNCTION: NAVIGATION COMPROMISED";
+                HUDManager.Instance.globalNotificationText.text =
+                    "SHIP MALFUNCTION:\nROUTE TABLE OFFSET ERROR";
+
                 HUDManager.Instance.globalNotificationAnimator.SetTrigger("TriggerNotif");
-                HUDManager.Instance.UIAudio.PlayOneShot(HUDManager.Instance.radiationWarningAudio, 1f);
+                HUDManager.Instance.UIAudio.PlayOneShot(
+                    HUDManager.Instance.radiationWarningAudio,
+                    1f
+                );
 
                 // Update the screen to have scrambled text.
-                __instance.screenLevelDescription.text = "SHIP NAVIGATION MALFUNCTION\nOrbiting: Unknown\nPopulation: Unknown\nConditions: Unknown\nFauna: Unknown\nWeather: Unknown";
-                __instance.screenLevelVideoReel.enabled = false;
+                __instance.screenLevelDescription.text =
+                    "SHIP NAVIGATION MALFUNCTION\nOrbiting: Unknown\nWeather: Unknown\nBE ADVISED THE ROUTE TABLE WILL RECALIBRATE AFTER LANDING";
 
-                // For some reason just setting .enabled to false here didn't work, so we also undo the other stuff it sets
+                // Adopted this code from stromytuna's RouteRandom mod.
+                __instance.screenLevelVideoReel.enabled = false;
                 __instance.screenLevelVideoReel.clip = null;
                 __instance.screenLevelVideoReel.gameObject.SetActive(false);
                 __instance.screenLevelVideoReel.Stop();
+
+                // Create sparks on the lever for the navigation malfunction.
+                StartMatchLever leverDevice =
+                    UnityEngine.Object.FindObjectOfType<StartMatchLever>();
+
+                if (leverDevice == null)
+                {
+                    Plugin.logger.LogError("Failed to find Terminal object.");
+
+                    return;
+                }
+
+                // Instantiate the new sparks object on top of the lever.
+                State.sparksNavigation = Assets.SpawnPrefab(
+                    "sparks",
+                    leverDevice.transform.position
+                );
+
+                // Assign the sparks to the map device so that they move together.
+                State.sparksNavigation.transform.SetParent(leverDevice.transform, true);
             }
-		}
-	}
+            else
+            {
+                // If there is no malfunction, destroy the sparks objects again.
+                if (State.sparksNavigation != null)
+                {
+                    UnityEngine.Object.Destroy(State.sparksNavigation);
+                }
+            }
+        }
+
+        [HarmonyPostfix]
+        [HarmonyPatch("Start")]
+        public static void Reset(StartOfRound __instance)
+        {
+            // Reset/setup the tracking of the malfunction states and objects.
+            State.Reset();
+        }
+    }
 }