+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 
 # Changelog #
 
+## 1.1.0 ##
+
+- Navigation Malfunction
+    + Improve notification text
+    + Improve navigation map text
+    + Add sparks to lever while active
+    + Block the user from choosing a different moon until malfunction ends
+
 ## 1.0.0 ##
 
 - Initial release with basic functionality